+++ Modules/xStorage/Examples/Resources/xOpticalDiskDriveLetter/1-xOpticalDiskDriveLetter_SetDriveLetter.ps1
@@ -1,6 +1,7 @@
 <#
     .EXAMPLE
-        This configuration will set the drive letter of the optical disk drive to 'Z'.
+        This configuration will set the drive letter of the first
+        optical disk drive in the system to 'Z'.
 #>
 Configuration Example
 {
@@ -9,10 +10,10 @@ Configuration Example
 
     Node localhost
     {
-        xOpticalDiskDriveLetter MapOpticalDiskToZ
+        xOpticalDiskDriveLetter SetFirstOpticalDiskDriveLetterToZ
         {
-            IsSingleInstance = 'Yes'
-            DriveLetter      = 'Z'
+            DiskId      = 1
+            DriveLetter = 'Z'
         }
     }
 }