+++ Tests/Integration/MSFT_xMountImage_ISO.Integration.Tests.ps1
@@ -35,7 +35,7 @@ try
     if (-not (Test-Path -Path $ISOPath))
     {
         Write-Verbose -Message "$($script:DSCResourceName) integration tests cannot be run because the ISO File '$ISOPath' is not available." -Verbose
-        Return
+        return
     } # if
 
     # Get a spare drive letter