+++ Tests/Integration/MSFT_xOpticalDiskDriveLetter.config.ps1
@@ -1,9 +1,20 @@
 configuration MSFT_xOpticalDiskDriveLetter_config {
     Import-DSCResource -ModuleName xStorage
     node localhost {
-        xOpticalDiskDriveLetter Integration_Test {
-            IsSingleInstance = 'Yes'
-            DriveLetter      = $Node.DriveLetter
+        if ($Node.Ensure)
+        {
+            xOpticalDiskDriveLetter Integration_Test {
+                DiskId      = $Node.DiskId
+                DriveLetter = $Node.DriveLetter
+                Ensure      = $Node.Ensure
+            }
+        }
+        else
+        {
+            xOpticalDiskDriveLetter Integration_Test {
+                DiskId      = $Node.DiskId
+                DriveLetter = $Node.DriveLetter
+            }
         }
     }
 }