+++ Tests/Unit/MSFT_xOpticalDiskDriveLetter.Tests.ps1
@@ -28,6 +28,8 @@ try
     # (non-exported) code of a Script Module.
     InModuleScope $script:DSCResourceName {
         $script:testDriveLetter = 'X:'
+        $script:testDriveLetterNoVolume = 'Volume{47b90a5d-f340-11e7-80fd-806e6f6e6963}'
+        $script:testVolumeDeviceId = '"\\?\$($script:testDriveLetterNoVolume)\"'
 
         $script:mockedNoOpticalDrive = $null
 
@@ -38,37 +40,51 @@ try
             Id       = $script:testDriveLetter
         }
 
-        $script:mockedVolume = [pscustomobject] @{
-            DriveLetter = $script:testDriveLetter
-            DriveType   = 5
+        $script:mockedOpticalDriveNoDriveLetter = [pscustomobject] @{
+            Drive    = $script:testDriveLetterNoVolume
+            Caption  = 'Microsoft Virtual DVD-ROM'
+            DeviceID = 'SCSI\CDROM&VEN_MSFT&PROD_VIRTUAL_DVD-ROM\000006'
+            Id       = $script:testDriveLetterNoVolume
         }
+
+        $script:mockedOpticalDriveMultiDisks = @(
+            $script:mockedOpticalDriveNoDriveLetter
+            $script:mockedOpticalDrive
+        )
 
         $script:mockedWrongLetterOpticalDrive = [pscustomobject] @{
             Drive    = 'W:'
             Caption  = 'Microsoft Virtual DVD-ROM'
             DeviceID = 'SCSI\CDROM&VEN_MSFT&PROD_VIRTUAL_DVD-ROM\000006'
-        }
-
-        $script:mockedWrongVolume = [pscustomobject] @{
-            IsSingleInstance = 'Yes'
-            DriveLetter      = 'W:'
-        }
-
-        $script:mockedVolumeNotOpticalDrive = [pscustomobject] @{
-            DriveLetter = $script:testDriveLetter
-            DriveType   = 3
+            Id       = 'W:'
         }
 
         $script:mockedOpticalDriveISO = [pscustomobject] @{
             Drive    = 'I:'
             Caption  = 'Microsoft Virtual DVD-ROM'
             DeviceID = 'SCSI\CDROM&VEN_MSFT&PROD_VIRTUAL_DVD-ROM\2&1F4ADFFE&0&000002'
+            Id       = 'I:'
         }
 
         $script:mockedOpticalDriveIDE = [pscustomobject] @{
             Drive    = 'I:'
             Caption  = 'Msft Virtual CD/ROM ATA Device'
             DeviceID = 'IDE\CDROMMSFT_VIRTUAL_CD/ROM_____________________1.0_____\5&CFB56DE&0&1.0.0'
+            Id       = 'I:'
+        }
+
+        $script:mockedVolume = [pscustomobject] @{
+            DriveLetter = $script:testDriveLetter
+            DriveType   = 5
+            DeviceId    = '\\?\Volume{bba1802b-e7a1-11e3-824e-806e6f6e6963}\'
+            Id          = $script:testDriveLetter
+        }
+
+        $script:mockedWrongVolume = [pscustomobject] @{
+            DriveLetter = 'W:'
+            DriveType   = 5
+            DeviceId    = $script:testVolumeDeviceId
+            Id          = 'W:'
         }
 
         function Set-CimInstance
@@ -86,21 +102,23 @@ try
         }
 
         #region Function Get-TargetResource
-        Describe 'MSFT_xOpticalDiskDriveLetter\Get-TargetResource' {
-            Context 'Optical disk drive present with correct drive letter' {
+        Describe 'MSFT_xOpticalDiskDriveLetter\Get-OpticalDiskDriveLetter' {
+            Context 'Single optical disk drive present and assigned a drive letter' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDrive } `
+                    -MockWith {
+                    $script:mockedOpticalDrive
+                } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Get-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -115,6 +133,157 @@ try
                 }
             }
 
+            Context 'Single optical disk drive present and not assiged a drive letter' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedOpticalDriveNoDriveLetter
+                } `
+                    -Verifiable
+
+                It 'Should not throw an exception' {
+                    {
+                        $script:result = Get-TargetResource `
+                            -DiskId 1 `
+                            -Driveletter $script:testDriveLetter `
+                            -Verbose
+                    } | Should -Not -Throw
+                }
+
+                It "DriveLetter should be empty" {
+                    $script:result.DriveLetter | Should -Be ''
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+
+            Context 'Multiple optical disk drives present and second one is assigned a drive letter' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedOpticalDriveMultiDisks
+                } `
+                    -Verifiable
+
+                It 'Should not throw an exception' {
+                    {
+                        $script:result = Get-TargetResource `
+                            -DiskId 2 `
+                            -Driveletter $script:testDriveLetter `
+                            -Verbose
+                    } | Should -Not -Throw
+                }
+
+                It "DriveLetter should be $($script:testDriveLetter)" {
+                    $script:result.DriveLetter | Should -Be $script:testDriveLetter
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+
+            Context 'Single optical disk drive present but second disk is requested' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedOpticalDrive
+                } `
+                    -Verifiable
+
+                $errorRecord = Get-InvalidArgumentRecord `
+                    -Message ($LocalizedData.NoOpticalDiskDriveError -f 2) `
+                    -ArgumentName 'DiskId'
+
+                It 'Should throw expected exception' {
+                    {
+                        $script:result = Get-TargetResource `
+                            -DiskId 2 `
+                            -Driveletter $script:testDriveLetter `
+                            -Verbose
+                    } | Should -Throw $errorRecord
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+
+            Context 'Single optical disk drive present but is mounted with ISO' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedOpticalDriveISO
+                } `
+                    -Verifiable
+
+                It 'Should throw expected exception' {
+                    {
+                        $script:result = Get-TargetResource `
+                            -DiskId 1 `
+                            -Driveletter $script:testDriveLetter `
+                            -Verbose
+                    } | Should -Throw $errorRecord
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+        }
+        #endregion
+
+        #region Function Get-TargetResource
+        Describe 'MSFT_xOpticalDiskDriveLetter\Get-TargetResource' {
+            Context 'Optical disk drive present with correct drive letter' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedOpticalDrive
+                } `
+                    -Verifiable
+
+                It 'Should not throw an exception' {
+                    {
+                        $script:result = Get-TargetResource `
+                            -DiskId 1 `
+                            -Driveletter $script:testDriveLetter `
+                            -Verbose
+                    } | Should -Not -Throw
+                }
+
+                It "DriveLetter should be $($script:testDriveLetter)" {
+                    $script:result.DriveLetter | Should -Be $script:testDriveLetter
+                    $script:result.Ensure | Should -Be 'Present'
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+
             Context 'Optical disk drive present with incorrect drive letter' {
                 # verifiable (should be called) mocks
                 Mock `
@@ -122,13 +291,15 @@ try
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedWrongLetterOpticalDrive } `
+                    -MockWith {
+                    $script:mockedWrongLetterOpticalDrive
+                } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Get-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -136,6 +307,7 @@ try
 
                 It "DriveLetter should be $($script:testDriveLetter)" {
                     $script:result.DriveLetter | Should -Not -Be $script:testDriveLetter
+                    $script:result.Ensure | Should -Be 'Present'
                 }
 
                 It 'Should call all the Get mocks' {
@@ -150,13 +322,15 @@ try
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDriveIDE } `
+                    -MockWith {
+                    $script:mockedOpticalDriveIDE
+                } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Get-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -164,6 +338,7 @@ try
 
                 It "Should be DriveLetter $($script:testDriveLetter)" {
                     $script:result.DriveLetter | Should -Not -Be $script:testDriveLetter
+                    $script:result.Ensure | Should -Be 'Present'
                 }
 
                 It 'Should call all the Get mocks' {
@@ -178,20 +353,22 @@ try
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedNoOpticalDrive } `
+                    -MockWith {
+                    $script:mockedNoOpticalDrive
+                } `
                     -Verifiable
 
-                It 'Should not throw an exception' {
+                $errorRecord = Get-InvalidArgumentRecord `
+                    -Message ($LocalizedData.NoOpticalDiskDriveError -f 1) `
+                    -ArgumentName 'DiskId'
+
+                It 'Should throw expected exception' {
                     {
                         $script:result = Get-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
-                    } | Should -Not -Throw
-                }
-
-                It 'DriveLetter should be null' {
-                    $script:result.DriveLetter | Should -Be $null
+                    } | Should -Throw $errorRecord
                 }
 
                 It 'Should call all the Get mocks' {
@@ -203,20 +380,22 @@ try
 
         #region Function Set-TargetResource
         Describe 'MSFT_xOpticalDiskDriveLetter\Set-TargetResource' {
-            Context 'Optical disk drive with the correct drive letter' {
+            Context 'Optical disk drive exists with the correct drive letter' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDrive } `
+                    -MockWith {
+                    $script:mockedOpticalDrive
+                } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         Set-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -228,14 +407,16 @@ try
                 }
             }
 
-            Context 'Optical disk drive with the correct drive letter when Ensure is set to Absent' {
+            Context 'Optical disk drive exists with a drive letter when Ensure is set to Absent' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDrive } `
+                    -MockWith {
+                    $script:mockedOpticalDrive
+                } `
                     -Verifiable
 
                 Mock `
@@ -243,7 +424,9 @@ try
                     -ParameterFilter {
                     $ClassName -eq 'Win32_Volume'
                 } `
-                    -MockWith { $script:mockedVolume } `
+                    -MockWith {
+                    $script:mockedVolume
+                } `
                     -Verifiable
 
                 Mock `
@@ -253,7 +436,7 @@ try
                 It 'Should not throw an exception' {
                     {
                         Set-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Ensure 'Absent' `
                             -Verbose
@@ -266,14 +449,16 @@ try
                 }
             }
 
-            Context 'Optical disk with the wrong drive letter' {
+            Context 'Optical disk drive exists with the wrong drive letter' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedWrongLetterOpticalDrive } `
+                    -MockWith {
+                    $script:mockedWrongLetterOpticalDrive
+                } `
                     -Verifiable
 
                 Mock `
@@ -281,7 +466,9 @@ try
                     -ParameterFilter {
                     $ClassName -eq 'Win32_Volume'
                 } `
-                    -MockWith { $script:mockedWrongVolume } `
+                    -MockWith {
+                    $script:mockedWrongVolume
+                } `
                     -Verifiable
 
                 Mock `
@@ -291,7 +478,7 @@ try
                 It 'Should not throw an exception' {
                     {
                         Set-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -304,14 +491,16 @@ try
                 }
             }
 
-            Context 'IDE optical disk drive with the wrong drive letter' {
+            Context 'IDE optical disk drive exists with the wrong drive letter' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDriveide } `
+                    -MockWith {
+                    $script:mockedOpticalDriveIDE
+                } `
                     -Verifiable
 
                 Mock `
@@ -319,7 +508,9 @@ try
                     -ParameterFilter {
                     $ClassName -eq 'Win32_Volume'
                 } `
-                    -MockWith { $script:mockedWrongVolume } `
+                    -MockWith {
+                    $script:mockedWrongVolume
+                } `
                     -Verifiable
 
                 Mock `
@@ -329,7 +520,7 @@ try
                 It 'Should not throw an exception' {
                     {
                         Set-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -342,57 +533,55 @@ try
                 }
             }
 
-            # This resource does not change the drive letter of mounted ISO images.
-            Context 'Mounted ISO with the wrong drive letter' {
+            Context 'Optical disk drive not present' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDriveISO } `
+                    -MockWith {
+                    $script:mockedNoOpticalDrive
+                } `
                     -Verifiable
 
-                It 'Should not throw an exception' {
+                $errorRecord = Get-InvalidArgumentRecord `
+                    -Message ($LocalizedData.NoOpticalDiskDriveError -f 1) `
+                    -ArgumentName 'DiskId'
+
+                It 'Should throw expected exception' {
                     {
-                        Set-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                        $script:result = Set-TargetResource `
+                            -DiskId 1 `
                             -Driveletter $script:testDriveLetter `
                             -Verbose
-                    } | Should -Not -Throw
+                    } | Should -Throw $errorRecord
                 }
 
-                It 'Should call the correct mocks' {
+                It 'Should call all the Get mocks' {
                     Assert-VerifiableMock
-                    Assert-MockCalled -CommandName Get-CimInstance -Exactly -Times 1
                 }
             }
         }
         #endregion
 
         Describe 'MSFT_xOpticalDiskDriveLetter\Test-TargetResource' {
-            Context 'Drive letter is a valid optical disk drive' {
+            Context 'Optical drive exists and is assigned expected drive letter' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDrive } `
-                    -Verifiable
-
-                Mock `
-                    -CommandName Get-CimInstance  `
-                    -ParameterFilter {
-                    $ClassName -eq 'Win32_Volume'
+                    -MockWith {
+                    $script:mockedOpticalDrive
                 } `
-                    -MockWith { $script:mockedVolume } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Test-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -DriveLetter $script:testDriveLetter `
                             -Verbose
                     } | Should -Not -Throw
@@ -407,28 +596,22 @@ try
                 }
             }
 
-            Context 'Drive letter is a valid optical disk drive and $Ensure is set to Absent' {
+            Context 'Optical drive exists but is assigned a drive letter but should not be' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDrive } `
-                    -Verifiable
-
-                Mock `
-                    -CommandName Get-CimInstance  `
-                    -ParameterFilter {
-                    $ClassName -eq 'Win32_Volume'
+                    -MockWith {
+                    $script:mockedOpticalDrive
                 } `
-                    -MockWith { $script:mockedVolume } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Test-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -DriveLetter $script:testDriveLetter `
                             -Ensure 'Absent' `
                             -Verbose
@@ -444,30 +627,64 @@ try
                 }
             }
 
-            Context 'There is no optical disk drive' {
+            Context 'The drive letter already exists on a volume that is not an optical disk drive' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedNoOpticalDrive } `
+                    -MockWith {
+                    $script:mockedWrongLetterOpticalDrive
+                } `
                     -Verifiable
 
                 Mock `
-                    -CommandName Get-CimInstance  `
+                    -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_Volume'
                 } `
-                    -MockWith { $script:mockedWrongVolume } `
+                    -MockWith {
+                    $script:mockedVolume
+                } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Test-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -DriveLetter $script:testDriveLetter `
-                            -Ensure 'Present' `
+                            -Verbose
+                    } | Should -Not -Throw
+                }
+
+                It 'Should return $true' {
+                    $script:result | Should -Be $true
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+
+            Context 'The optical drive is assigned a drive letter but should not be' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedOpticalDrive
+                } `
+                    -Verifiable
+
+                It 'Should not throw an exception' {
+                    {
+                        $script:result = Test-TargetResource `
+                            -DiskId 1 `
+                            -DriveLetter $script:testDriveLetter `
+                            -Ensure 'Absent' `
                             -Verbose
                     } | Should -Not -Throw
                 }
@@ -481,36 +698,60 @@ try
                 }
             }
 
-            Context 'The drive letter already exists on a volume that is not a optical disk drive' {
+            Context 'The optical drive is not assigned a drive letter and should not be' {
                 # verifiable (should be called) mocks
                 Mock `
                     -CommandName Get-CimInstance `
                     -ParameterFilter {
                     $ClassName -eq 'Win32_CDROMDrive'
                 } `
-                    -MockWith { $script:mockedOpticalDrive } `
-                    -Verifiable
-
-                Mock `
-                    -CommandName Get-CimInstance  `
-                    -ParameterFilter {
-                    $ClassName -eq 'Win32_Volume'
+                    -MockWith {
+                    $script:mockedOpticalDriveNoDriveLetter
                 } `
-                    -MockWith { $script:mockedVolumeNotOpticalDrive } `
                     -Verifiable
 
                 It 'Should not throw an exception' {
                     {
                         $script:result = Test-TargetResource `
-                            -IsSingleInstance 'Yes' `
+                            -DiskId 1 `
                             -DriveLetter $script:testDriveLetter `
-                            -Ensure 'Present' `
+                            -Ensure 'Absent' `
                             -Verbose
                     } | Should -Not -Throw
                 }
 
-                It 'Should return $false' {
-                    $script:result | Should -Be $false
+                It 'Should return $true' {
+                    $script:result | Should -Be $true
+                }
+
+                It 'Should call all the Get mocks' {
+                    Assert-VerifiableMock
+                }
+            }
+
+            Context 'Optical disk drive not present' {
+                # verifiable (should be called) mocks
+                Mock `
+                    -CommandName Get-CimInstance `
+                    -ParameterFilter {
+                    $ClassName -eq 'Win32_CDROMDrive'
+                } `
+                    -MockWith {
+                    $script:mockedNoOpticalDrive
+                } `
+                    -Verifiable
+
+                $errorRecord = Get-InvalidArgumentRecord `
+                    -Message ($LocalizedData.NoOpticalDiskDriveError -f 1) `
+                    -ArgumentName 'DiskId'
+
+                It 'Should throw expected exception' {
+                    {
+                        $script:result = Test-TargetResource `
+                            -DiskId 1 `
+                            -Driveletter $script:testDriveLetter `
+                            -Verbose
+                    } | Should -Throw $errorRecord
                 }
 
                 It 'Should call all the Get mocks' {
@@ -518,7 +759,6 @@ try
                 }
             }
         }
-
         #endregion
     }
 }